requirements.txt
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.8.1
Automat==0.6.0
certifi==2017.7.27.1
channels==1.1.6
chardet==3.0.4
constantly==15.1.0
coverage==4.0.3
daphne==1.3.0
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.4
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
msgpack-python==0.4.8
psycopg2==2.7.3.1
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.10.0
Twisted==17.5.0
txaio==2.8.1
urllib3==1.22
zope.interface==4.4.2
